Move binaryop files to legacy
