+++ conda/recipes/libcudf/meta.yaml
@@ -53,7 +53,7 @@ test:
     - test -f $PREFIX/include/cudf/utilities/legacy/type_dispatcher.hpp
     - test -f $PREFIX/include/cudf/utilities/legacy/wrapper_types.hpp
     - test -f $PREFIX/include/cudf/utilities/error.hpp
-    - test -f $PREFIX/include/cudf/binaryop.hpp
+    - test -f $PREFIX/include/cudf/legacy/binaryop.hpp
     - test -f $PREFIX/include/cudf/convert_types.h
     - test -f $PREFIX/include/cudf/copying.hpp
     - test -f $PREFIX/include/cudf/legacy/copying.hpp
