+++ cpp/src/binaryop/legacy/binaryop.cpp
