+++ cpp/tests/binaryop/legacy/unit/binop-verify-input-test.cpp
@@ -20,7 +20,7 @@
 #include <tests/utilities/legacy/column_wrapper.cuh>
 #include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
-#include <cudf/binaryop.hpp>
+#include <cudf/legacy/binaryop.hpp>
 #include <gtest/gtest.h>
 #include <vector>
 #include <numeric>
