+++ cpp/tests/wrappers/timestamps_test.cu
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <cudf/binaryop.hpp>
+#include <cudf/legacy/binaryop.hpp>
 #include <cudf/column/column_factories.hpp>
 #include <cudf/column/column_device_view.cuh>
